chore(build): refactor
