+++ src/asian_options.py
@@ -2,9 +2,9 @@ from dataclasses import dataclass
 from datetime import date
 from math import exp, log, sqrt
 
-from black_scholes import calculate_black_scholes as bs
-from enums import OptionType
-from helpers import averaging_period, time_to_maturity
+from .black_scholes import calculate_black_scholes as bs
+from .enums import OptionType
+from .helpers import averaging_period, time_to_maturity
 
 
 @dataclass
@@ -18,9 +18,10 @@ def floating_ratio(pricing_date: date, start_date: date, end_date: date) -> floa
     tell me how much of the pricing period is still open, and how much of
     it has priced out already.
     """
+    unused_variable = 1
     if pricing_date > end_date:
         return 0
-    if start_date <     pricing_date:
+    if start_date < pricing_date:
         total = (end_date - start_date).days + 1
         open_part = (end_date - pricing_date).days + 1
